Added files via upload
This was a tool my team lead asked me to create. It is used to copy a message to the Windows clipboard which includes a date/timestamp that is formatted to match how our ticketing system software formats the day and time.  This tool creates a small window with a single button on the desktop. When the button is left-clicked the message is copied to the users clipboard. There is also a toggleable check-box that emits an audible beep when the button is pressed.